requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
idna==3.3
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.20.0
requests==2.28.1
sqlparse==0.4.2
update-checker==0.18.0
urllib3==1.26.11
websocket-client==1.3.3
asgiref==3.5.2
boto3==1.24.24
botocore==1.27.24
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
decorator==4.4.2
dill==0.3.5.1
Django==4.0.6
greenlet==1.1.2
gTTS==2.2.4
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
jmespath==1.0.1
loguru==0.6.0
lxml==4.9.1
moviepy==1.0.3
multiprocess==0.70.13
numpy==1.23.1
pathos==0.2.9
Pillow==9.2.0
playwright==1.23.0
pox==0.3.1
ppft==1.7.6.5
praw==7.6.0
prawcore==2.3.0
proglog==0.1.10
pyee==8.1.0
PyExecJS==1.5.1
Pygments==2.12.0
python-dateutil==2.8.2
pytube==12.1.0
requests==2.28.1
rich==12.5.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.0
translators==5.3.1
update-checker==0.18.0
websocket-client==1.3.3
websockets==10.1
